Optimum path sequence-hmm-viterbi
After POS(Parts Of Speech) tagging the sentences in a text corpus using nltk, I decided to use viterbi algorithm using Hidden Markov Model to find optimum path for every sentence in my text. Example:

[('Primavera-Sound-Festival', 'NN'), ('in', 'IN'), ('May/June', 'NNP'), ('this', 'DT'), ('popular', 'JJ'), ('alternative', 'JJ'), ('rock', 'NN'), ('and', 'CC'), ('dance', 'NN'), ('festival', 'NN'), ('takes', 'VBZ'), ('place', 'NN'), ('in', 'IN'), ('late', 'JJ'), ('spring/early', 'JJ'), ('summer', 'NN'), ('each', 'DT'), ('year', 'NN'), ('at', 'IN'), ('Parc-del-Forum', 'NNP'), ('easily', 'RB'), ('accessible', 'JJ'), ('on', 'IN'), ('the', 'DT'), ('Northern', 'NNP'),('edge', 'NN'), ('of', 'IN'), ('the', 'DT'), ('city', 'NN'), ('extensive', 'JJ'), ('line', 'NN'), ('ups', 'NNS'), ('with', 'IN'), ('swathes', 'NNS'), ('of', 'IN'),('fashionable', 'JJ'), ('indie', 'NN'), ('and', 'CC'), ('dance', 'VB'), ('a', 'DT'), ('real', 'JJ'), ('music', 'NN'), ('lovers', 'NNS'), ('event', 'NN'), ('that', 'WDT'), ('attracts', 'VBZ'), ('a', 'DT'), ('glamorous', 'JJ'), ('international', 'JJ'), ('crowd', 'NN'), ('.', '.')]

In order to pick this optimum sequence of tags for the sentence the algorithm goes backward in finding the previous tag and the probabilities of tags that can possibly follow that previous one and check for the value that the previous one carries for the current one. Multiply that probability value with that of the current tag and store this value as state transition value, meaning: what's the value of transitioning from previous tag to the current tag. And this calculation continues till the end of a sentence. Finally the algorithm picks a path that carries maximum probability for the sentence. All of this would make a lot of sentence with a possibility of having at least one tag for a word in a sentence. Ought to illustrate the idea with an example: from my dataset:
The word ‘carnival’ gets accompanied with two tags and their probabilities as below:

'carnival': [('NN', 0.8), ('VBP', 0.2)]

Now the flow goes in searching for the previous tag that had appeared with a word preceding ‘carnival’. The previous word is ‘countries’ and so its tag combo is 'countries': [('NNS', 1.0)]. Now it checks for the possible tags that might follow the tag ‘NNS’ and the list appears below:

'NNS': [('.', 0.0),
   (':', 0.0),
   ('CC', 0.1),
   ('DT', 0.0),
   ('IN', 0.3),
   ('JJ', 0.0),
   ('JJR', 0.0),
   ('MD', 0.0),
   ('NN', 0.0),
   ('NNP', 0.0),
   ('NNS', 0.1),
   ('PRP', 0.0),
   ('RB', 0.0),
   ('RP', 0.0),
   ('TO', 0.0),
   ('VBD', 0.0),
   ('VBG', 0.0),
   ('VBN', 0.0),
   ('VBP', 0.2),
   ('VBZ', 0.0),
   ('WDT', 0.0),
   ('WRB', 0.0)]

The algorithm looks for the value for both the current tags that are associated with the current word ‘carnival’. For ‘NN’ the value appears as 0.0 in the list of the previous tag ‘NNS’ and so multiplying gives a value 0. But for ‘VBP’ the value appears as 0.2 and so multiplying that with the current value 0.2 for VBP I get 0.04. Thus the maximum of the two probabilities is chosen and thus ‘VBP’ becomes the appropriate tag for the word ‘carnival’. Although it doesn’t make sense to tag ‘carnival’ as a verb, the previous word was actually just ‘countries’ it was actually ‘countries’’- a possessive noun. Because the algorithm failed to capture that the current word has received the ‘VBP’ tag terming the word as a verb. Had the algorithm captured the possessive noun properly then instead of searching the tag ‘NNS’-which means plural proper noun it would have searched the tag ‘PRP$’ and its list of tags that might follow is:
'PRP$': [('JJ', 0.1), ('NN', 0.5), ('NNS', 0.3), ('VBN', 0.1)]
With this option it’s clear that ‘NN’ carries a lot of weight and so multiplying the value with 0.8 of the current ‘NN’ tag gives the maximum value and therefore ‘NN’ would have been chosen for the ‘carnival’.