Phase II: NMF to cluster similar events across cities!
Phase II of EmoU.club uses the description of events from every city and performs Non-Negative Matrix Factorization to identify clusters of similar events/happenings across cities. This was done to identify any future potential of the project. It turned out to be quite interesting clusters to transform to a personalized recommendation engine in the future.