Extracting Named Entities
There are 5 different regex patterns illustrated in this commit to extract maximum information about different topics, say: events, dates, significant words describing power with regards to a specific region, etc… Have got patterns to fit the context of my dataset (the one used in EmoU.club) and have also tried a few patterns on some random Wikipedia text. The only drawback with regex patterns is that they are very constrained in terms of implementation and so cannot be extended widely across different applications.